separates out provider list into store
